+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 0.21.0 (TBD)
 
+- Switched Falcon512 implementation to use `fn-dsa` crate ([#774](https://github.com/0xMiden/crypto/pull/774)).
 - [BREAKING] Removed `p3-compat` and `winter-compat` features ([#745](https://github.com/0xMiden/crypto/pull/745)).
 - Use more idiomatic Plonky3 APIs ([#743](https://github.com/0xMiden/crypto/pull/743)).
 - Make concurrent feature interact with plonky3's parallel features, replace homegrown iterator macros with p3-maybe-rayon ([#749](https://github.com/0xMiden/crypto/pull/749)).